requirements.txt
streamlit
pymupdf
pytesseract
pdf2docx
pillow
python-docx